feat: Add option to download pure lyrics
Introduces a new feature allowing users to download the generated LRC content as plain text, without timestamps. This caters to users who may want to extract only the lyric text for other purposes.

The default download extension has also been changed to `.txt` to better reflect the new functionality and to simplify the user experience for pure lyric downloads.

Updates the UI with a new icon for the "download pure lyrics" action and removes a redundant GH-Pages link from the README.